backend/requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
django-environ==0.11.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2